Refactor get_plot to use as_vb_dataframe handler
This adds a new internal function intended to work on common VegBank API
JSON responses containing a top-level "error" key (with informative
message) in the case of an error, and otherwise containing a top-level
"data" object (coercible to a data frame), optionally with a
corresponding top-level "count" field indicating the record count. When
the response contains data records, these are returned in a dataframe
with canonicalized names (by default), whereas error conditions or
zero-record cases are handled appropriately.