Add wrapper for get_observation_details endpoint
Note that this API call returns a more complicated JSON response than
the other main GET resource endpoints. If the provided accession code
matches a plot observation in the DB, the response is a JSON list that
contains one element. This element has keys "taxa" and "communities"
that themselves each correspond to a list of table records (for taxon
observations and community classifications, respectively), while all
other keys have a single atomic value and collectively correspond to
fields of the plot observation record itself. The R wrapper function
handles all of this, returning a list of three dataframes
(plot_observation, taxa, and communities).