Remove API error response handling outside of `send`
The latest version of the API only returns JSON responses with a top
level "error" key when there's a true error, and it now sends an HTTP
error status code in those cases. This is already handled in the
internal `send()` function, which currently stops execution with an R
error in the event of an HTTP error (consistent with httr2 behavior).
Hence, no reason to check for this in the higher level package functions
that call and/or process the output of `send()`.